+++ src/eopf_stac/sentinel3/stac.py
@@ -11,6 +11,8 @@ from eopf_stac.common.constants import (
     SENTINEL_LICENSE,
     SUPPORTED_S3_OLCI_L1_PRODUCT_TYPES,
     SUPPORTED_S3_OLCI_L2_PRODUCT_TYPES,
+    SUPPORTED_S3_SLSTR_L1_PRODUCT_TYPES,
+    SUPPORTED_S3_SLSTR_L2_LST_PRODUCT_TYPE,
     THUMBNAIL_ASSET,
 )
 from eopf_stac.common.stac import (
@@ -31,6 +33,10 @@ from eopf_stac.sentinel3.constants import (
     OLCI_L2_ASSETS,
     OLCI_L2_ASSETS_KEY_TO_PATH,
     SENTINEL3_METADATA,
+    SLSTR_L1_ASSETS,
+    SLSTR_L1_ASSETS_KEY_TO_PATH,
+    SLSTR_L2_LST_ASSETS,
+    SLSTR_L2_LST_ASSETS_KEY_TO_PATH,
 )
 
 logger = logging.getLogger(__name__)
@@ -42,7 +48,7 @@ def create_collection(collection_metadata: dict, thumbnail_href: str) -> pystac.
         "constellation": [mission_metadata.get("constellation")],
         "platform": mission_metadata.get("platforms"),
         "instruments": [collection_metadata.get("instrument")],
-        "gsd": [collection_metadata.get("gsd")],
+        "gsd": collection_metadata.get("gsd"),
         "processing:level": [collection_metadata.get("processing_level")],
         "product:type": [collection_metadata.get("product_type")],
     }
@@ -152,6 +158,12 @@ def create_item(metadata: dict, product_type: str, asset_href_prefix: str) -> py
     elif product_type in SUPPORTED_S3_OLCI_L2_PRODUCT_TYPES:
         asset_defintions = OLCI_L2_ASSETS
         asset_path_lookups = OLCI_L2_ASSETS_KEY_TO_PATH
+    elif product_type in SUPPORTED_S3_SLSTR_L1_PRODUCT_TYPES:
+        asset_defintions = SLSTR_L1_ASSETS
+        asset_path_lookups = SLSTR_L1_ASSETS_KEY_TO_PATH
+    elif product_type in SUPPORTED_S3_SLSTR_L2_LST_PRODUCT_TYPE:
+        asset_defintions = SLSTR_L2_LST_ASSETS
+        asset_path_lookups = SLSTR_L2_LST_ASSETS_KEY_TO_PATH
     else:
         raise ValueError(f"Unsupported Sentinel-3 product type '{product_type}'")